feat(more-commands): Added binaries to the gitignore, Refactored build steps with Cmake, Seperated classes into hpp/cpp where neeeded.
