+++ .gitignore
@@ -1,0 +1,9 @@
+#build folder
+/build
+
+# Built binaries
+wiff
+/wiff/usr/bin/wiff
+
+# Debian packages
+*.deb
